feat: add auth setup and dependency
