+++ playwright/.auth/user.json
@@ -1,0 +1,75 @@
+{
+  "cookies": [
+    {
+      "name": "next-auth.csrf-token",
+      "value": "564fe22e85eb7fb4e0ff592987ebf56574bcd060f21e8c06e524bc4ba1b067ee%7C6e09421ced70c6772dcbd0aa564cbf451fd2dfa1ec874e802b84d8f61bdb7af5",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "next-auth.callback-url",
+      "value": "http%3A%2F%2Flocalhost%3A3000%2F",
+      "domain": "localhost",
+      "path": "/",
+      "expires": -1,
+      "httpOnly": true,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "__Host-zitadel.login.csrf",
+      "value": "MTY4NTU1ODIxMXxJbmRLYVVkdU4zRTBSREExYjFkM2NWTlFXUzl0VkVGcVdWcG5Xak5sTXpkUlFUQlRUMFpUYmtOcmEyczlJZ289fJ-rsBei-vkTKvYxqTwkkEeu73ZkC7ZYjKVHQilTTZ_2",
+      "domain": "cas-fee-advanced-ocvdad.zitadel.cloud",
+      "path": "/",
+      "expires": 1685601411.716016,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "mode",
+      "value": "auto-light",
+      "domain": "cas-fee-advanced-ocvdad.zitadel.cloud",
+      "path": "/",
+      "expires": 1717094212,
+      "httpOnly": false,
+      "secure": false,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "__Secure-zitadel.useragent",
+      "value": "MTY4NTU1ODIxNHxib1M1eUxSOGhsWllfZ0hfb1ZjZm5WcnNYTS1tQW51SHE5a1JOSHl6dy1FSXU3NFh6RUlZYlpZX1Z6VkFnRUVzVkR5a2xlRjJnTTE4NE1hY0trMnRCak1lTlkwUDNkZz18yQj61zqsOHoxkmHNpbXPHgRodtUJxxbqY7x4RQrkd-g=",
+      "domain": ".cas-fee-advanced-ocvdad.zitadel.cloud",
+      "path": "/",
+      "expires": 1717094214.579312,
+      "httpOnly": true,
+      "secure": true,
+      "sameSite": "Lax"
+    },
+    {
+      "name": "next-auth.session-token",
+      "value": "eyJhbGciOiJkaXIiLCJlbmMiOiJBMjU2R0NNIn0..Qo3YWI0QnasrzTHZ.Z7Xsev8gc6pfjU_PIBWxL1V91RBeOn8uBgZrlT1ZLY4s8rWFCpGO4NkXlV7uSmRnZzeaMcgDRliTo0SPzS6TpSYiA3-FGUU18e-h-J4w2OTPfjajoERRxb21ZHn802_sz91q0kE-ATaxzodOwlbwqWDC2EOrO_jEGqUiREPUalYJNlNdTrgDgfdZXvaC3BPk9DOrLDx6aesyj4s-3I7tVeogaDwjUbJVY_KnNGRJuBMBpcNRmDKUn27hUuy4BllASJRJrQMz4aTaWC02hPd4VpHAgaAB4ZX4kvmmMvKMkWs4fUHiVrkxaTBXaH9jrCd7yG0FH63dcILpdk1hYUQBggvJF8F2LV9SgobZHi7UGmexVRl90lAZ27H9w2hbru8x7tW91j4WLPaouAGZpIXMmdSyXLlngSVqOPbbYjEXH7YHQxCBP6eP8XnRGT-alvdWWUjTkYTMQLxpz--TzA6TkFmGHO-xdFlC2K7lz9yw18pSKZWrrE0BUBWCkKtTUbGQq3yke-qBUquK2R_2q-iJaWAh4HdN3512cXgjdkJ8AGvFXGvMCayemDD40_TMszwlNusMHn4SZsihv32I_nAjOaVCBIeVfxvor4PVzhkJQ4ET7RolTyCPKqvpiss.9MGF4Lsmx---LPEn-Xr3rA",
+      "domain": "localhost",
+      "path": "/",
+      "expires": 1685601415.509628,
+      "httpOnly": true,
+      "secure": false,
+      "sameSite": "Lax"
+    }
+  ],
+  "origins": [
+    {
+      "origin": "http://localhost:3000",
+      "localStorage": [
+        {
+          "name": "nextauth.message",
+          "value": "{\"event\":\"session\",\"data\":{\"trigger\":\"getSession\"},\"timestamp\":1685558216}"
+        }
+      ]
+    }
+  ]
+}
